+++ portal/src/app/common/NavBar.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router'
 import NavMenuButton from './NavMenuButton.jsx'
-import {ALL_NAV} from '../store/Navigation.js'
+import {ALL_NAV} from '../models/Navigation.js'
 require('./NavBar.css');
 
 class NavBar extends React.Component {
