+++ portal/src/app/models/AppSettings.js
