+++ portal/src/app/models/ColorBlender.js
