+++ portal/src/app/models/CustomResultSearch.js
