+++ portal/src/app/models/DnaSequences.js
