+++ portal/src/app/models/HeatMapData.js
