+++ portal/src/app/models/PredictionDetail.js
