+++ portal/src/app/models/PredictionSettings.js
