+++ portal/src/app/models/StreamValue.js
